Refactor Feedback Log component
